requirements.txt
pbr>=0.6,!=0.7,<1.0
python-keystoneclient>=0.10.0
